+++ software/EveryDayCalendar/src/modes/clocks.h
@@ -1,0 +1,9 @@
+#ifndef __CLOCK_H__
+#define __CLOCK_H__
+
+#include <Arduino.h>
+
+void modeBinaryClock( time_t now);
+void modeEPOCHClock( time_t now);
+
+#endif // __CLOCK_H__